🎉First commit: Shione was born!🥳
In this release, we've released the first version v0.0.1 of the Shione project as a preview. This is somewhat like a sister project to Shigure-Bot, as it's base on that, and the developers are the same :)

To be honest, we're still at a very early stage now. But I believe that my passion can last at least for a while...

Change log:

- Introduce basic bot framework
- Add partial support for OneBot-V11 Events
- Add a basic plugin system
- Add a plugin: RepeaterPlugin